dodano potwierdzenie usunięcia hasłem admina

+++ Obsługa_serwisu/panelAdmin.cs
@@ -9,6 +9,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using Microsoft.VisualBasic;
 
 namespace Obsługa_serwisu
 {
@@ -25,7 +26,7 @@ namespace Obsługa_serwisu
             PoloczenieMySQL polMySQL = new PoloczenieMySQL();
             MySqlConnection polaczenie = polMySQL.polacz();
             string komenda = "INSERT INTO " + "pracownicy" + "(haslo) "
-            + "VALUES ('" +  textBoxPass.Text + "')  WHERE login = '"+textBoxLogin.Text+"';";
+            + "VALUES ('" +  textBoxPass.Text + "')  WHERE login = "+textBoxLogin.Text+";";
             MySqlCommand pytanie = new MySqlCommand(komenda, polaczenie);
             MySqlDataReader wynik;
             MessageBox.Show("Klient dodany");
@@ -34,6 +35,7 @@ namespace Obsługa_serwisu
             polMySQL.zamknij();
         }
 
+        
         private void button3_Click(object sender, EventArgs e)
         {
 
@@ -41,16 +43,27 @@ namespace Obsługa_serwisu
                 MessageBoxButtons.YesNoCancel, MessageBoxIcon.Question);
             if (dial == DialogResult.Yes)
             {
-                PoloczenieMySQL polMySQL = new PoloczenieMySQL();
-                MySqlConnection polaczenie = polMySQL.polacz();
-                string komenda = "DELETE FROM pracownicy WHERE login = '"+textBoxLoginDel.Text+"';";
-                MySqlCommand pytanie = new MySqlCommand(komenda, polaczenie);
-                MySqlDataReader wynik;
-                MessageBox.Show("Pracownik usunięty");
+                string input = Interaction.InputBox("Podaj hasło administratora", "Potwierdzenie usunięcia", "Default", -1, -1);
+                if (input.ToString() == "admin")
+                {
+                    PoloczenieMySQL polMySQL = new PoloczenieMySQL();
+                    MySqlConnection polaczenie = polMySQL.polacz();
+                    string komenda = "DELETE FROM pracownicy WHERE login = '" + textBoxLoginDel.Text + "';";
+                    MySqlCommand pytanie = new MySqlCommand(komenda, polaczenie);
+                    MySqlDataReader wynik;
+                    MessageBox.Show("Pracownik usunięty");
 
-                wynik = pytanie.ExecuteReader();
-                polMySQL.zamknij();
-            }
+                    wynik = pytanie.ExecuteReader();
+                    polMySQL.zamknij();
+                    textBoxLoginDel.Text = "";
+                }
+                else
+                    if (input.ToString() != "admin")
+                {
+                    MessageBox.Show("Błędne hasło");
+                    string input1 = Interaction.InputBox("Podaj hasło administratora", "Potwierdzenie usunięcia", "Default", -1, -1);
+                }
+                }
             else
                 if (dial == DialogResult.No)
             {
@@ -69,7 +82,7 @@ namespace Obsługa_serwisu
             MySqlConnection polaczenie = polMySQL.polacz();
             string komenda = "INSERT INTO " + "pracownicy" + "(Imie,Nazwisko,Adres,login,haslo) "
             + "VALUES ('" + textBoxNameAdd.Text + "', '" + textBoxSuAdd.Text + "','" 
-            + textBoxAdrAdd.Text + "','" + textBox5LoginAdd.Text + "','"+textBoxPassAdd+ "');" +
+            + textBoxAdrAdd.Text + "','" + textBox5LoginAdd.Text + "','"+textBoxPassAdd.Text+ "');" +
             " INSERT INTO " + "dzialy" + "(Nazwa_dzialu) " + "VALUES ('" + textBoxDzAdd.Text + "');";
           
             MySqlCommand pytanie = new MySqlCommand(komenda, polaczenie);
